Restructure compiler
