+++ packages/solid-marked/compiler/types.ts
@@ -1,0 +1,15 @@
+import GithubSlugger from 'github-slugger';
+import { SourceNode } from 'source-map';
+
+export interface Options {
+  mdxImportSource?: string;
+  noDynamicComponents?: boolean | 'only-mdx';
+}
+
+export interface StateContext {
+  source: string;
+  imports: SourceNode[];
+  options: Options;
+  frontmatter?: SourceNode;
+  slugger: GithubSlugger;
+}
