+++ packages/solid-marked/test/index.test.ts
@@ -58,8 +58,7 @@ test('should support InlineCode', () => {
 });
 test('should support Break', () => {
   expect(testCompile(`
-foo··
-bar
+a  \nb
 `)).toMatchSnapshot();
 });
 test('should support Link', () => {
